Split ITests in two parts: one for slaves and one with a single Jenkins instance.
